+++ breeze-dsl/src/main/kotlin/com/windea/breezeframework/dsl/graph/FlowChartDsl.kt
@@ -1,0 +1,12 @@
+package com.windea.breezeframework.dsl.graph
+
+import com.windea.breezeframework.core.annotations.api.*
+
+//TODO
+
+//REGION top annotations and interfaces
+
+annotation class FlowDsl
+
+@ReferenceApi("[Github](//https://github.com/adrai/flowchart.js)")
+class FlowChart