+++ breeze-dsl/src/main/kotlin/com/windea/breezeframework/dsl/graph/SequenceDiagramDsl.kt
@@ -1,0 +1,9 @@
+package com.windea.breezeframework.dsl.graph
+
+//TODO
+
+//REGION top annotations and interfaces
+
+annotation class SequenceDiagramDsl
+
+class SequenceDiagram